pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transparent-background-mcp"
version = "0.1.0"
description = "Local AI-powered background removal MCP server"
authors = [{name = "Joe Leaver", email = "joe@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["mcp", "background-removal", "ai", "computer-vision"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.2.0",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.3.0",
    "transparent-background>=1.3.4",
    "rembg>=2.0.67",
    "huggingface-hub>=0.20.0",
    "pydantic>=2.0.0",
    "psutil>=5.9.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
transparent-background-mcp-server = "transparent_background_mcp.server:cli_main"

[project.urls]
Homepage = "https://github.com/joeleaver/transparent-background-mcp"
Repository = "https://github.com/joeleaver/transparent-background-mcp"
Issues = "https://github.com/joeleaver/transparent-background-mcp/issues"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
